# Switch to json-iterator for v1 api.