This makes queries ~15% faster and cuts cpu
time spent on json encoding by ~40%.