@brian-brazil 2017-12-03
> Did you do any manual comparisons on how much speed boost this gives us?

See the PR description :)