@fabxc 2017-12-03
He, oops my bad. See my comment regarding streaming. Would be curious how much extra speed boost that could give.