@brian-brazil 2017-12-04
Streaming actually seems to increase usage ~10% compared to marshall, which is a little surprising.