@jacksontj 2017-12-21
There are larger performance gains in https://github.com/prometheus/common/pull/112