@brian-brazil 2017-12-21
Have you performed an end-to-end benchmark for a fair comparison?