@jacksontj 2017-12-21
@brian-brazil end-to-end meaning the http API response? The PR already includes data serializing an entire matrix response (which is pretty much end-to-end). I have pasted in my other issue on prom (https://github.com/prometheus/prometheus/issues/3601#issuecomment-353423635) with a complete end-to-end number (3m -> 16s).