@jacksontj 2017-12-21
@brian-brazil I've also submitted a PR (https://github.com/prometheus/prometheus/pull/3608) which has data on a full end-to-end API calls.