@brian-brazil 2018-02-08
I've worked a bit more on this:

```
Before:
BenchmarkRespond-4           100          16317782 ns/op         3020949 B/op      60009 allocs/op

With this PR:
BenchmarkRespond-4           500           3915364 ns/op          692664 B/op      20010 allocs/op

With this PR and improved WriteFloat64:
BenchmarkRespond-4           500           2975517 ns/op          213472 B/op         10 allocs/op

```

https://github.com/json-iterator/go/pull/234 is the change to improve json-iterator, if that gets in we're looking at a total improvement of ~5.5x on JSON encoding time and an elimination of per-Point allocations.
For very simple PromQL queries that should be a ~40% reduction in latency, as roughly equal time is currently spent on PromQL and encoding.