@jacksontj 2018-02-09
@brian-brazil I can't seem to find the code for the benchmark `BenchmarkRespond` that you referenced. Is that in another repository? Or to be added to the PR?