@jacksontj 2018-02-09
It seems that the structs involved have moved a decent amount since my last benchmarks. I've created a branch on my fork with an implementation of this in easyjson for comparison.

To start off, I'll show the comparison of baseline and this json iter one:

```
benchmark              old ns/op     new ns/op     delta
BenchmarkRespond-4     12054905      2277690       -81.11%

benchmark              old allocs     new allocs     delta
BenchmarkRespond-4     90021          20010          -77.77%

benchmark              old bytes     new bytes     delta
BenchmarkRespond-4     4629942       692666        -85.04%
```

Here are the compared results for baseline and easyjson: 
````
benchmark              old ns/op     new ns/op     delta
BenchmarkRespond-4     12054905      1785230       -85.19%

benchmark              old allocs     new allocs     delta
BenchmarkRespond-4     90021          10120          -88.76%

benchmark              old bytes     new bytes     delta
BenchmarkRespond-4     4629942       42369         -99.08%
```

And for simpler comparison here is the data comparing jsoniter to easyjson:
```
benchmark              old ns/op     new ns/op     delta
BenchmarkRespond-4     2277690       1785230       -21.62%

benchmark              old allocs     new allocs     delta
BenchmarkRespond-4     20010          10120          -49.43%

benchmark              old bytes     new bytes     delta
BenchmarkRespond-4     692666        42369         -93.88%
```

The benchmarks here show that there is ~21% performance improvement, and ~50% fewer allocs as well as ~93% fewer bytes allocated.

I then pulled down your jsoniter patch and compared easyjson to jsoniter (with the float improvements):
```
benchmark              old ns/op     new ns/op     delta
BenchmarkRespond-4     1785230       1552121       -13.06%

benchmark              old allocs     new allocs     delta
BenchmarkRespond-4     10120          10             -99.90%

benchmark              old bytes     new bytes     delta
BenchmarkRespond-4     42369         213474        +403.84%
```
With this wee see that it is ~13% faster. The tradeoff being that it allocates *significantly* more memory (~5x) albeit in fewer allocations. To be fair the easyjson implementation of mine here was done in ~30m, so there's definitely room for some improvement-- so at this point I'd say the performance between the 2 seems about the same (which is quite impressive!).

The main difference remaining between the 2 is that jsoniter is still doing a full marshal and then write (meaning the marshal can't be canceled). IMO this can still be a large concern, as some responses take quite a while to marshal out. That being said, speeding it up dramatically like this will definitely alleviate that concern (even if it doesn't solve it outright).