@brian-brazil 2018-02-09
@jacksontj I have discussed this with some of the other developers and while the gains of easyjson are nice, the maintenance overhead is just too high. json-iterator provides us basically with the same gains and no notable maintenance overhead.