@brian-brazil 2018-02-14
Upstream has taken my PR (and then improved upon it), so I've updated my PR to re-vendor. I think this is now ready to be taken in.