@brancz 2018-02-14
Do you think this is safe to land in 2.2? I think I would feel more comfortable with this being RCed as well.